separates client responsibility in files
